spinslide int to float

+++ PA7/src/object.cpp
@@ -1,6 +1,6 @@
 #include "object.h"
 #include <math.h> 
-extern int speedChange;
+extern float speedChange;
 
 Object::Object()
 {
